Fix webpack `graphql` misresolution by renaming `graphql.ts` module (#8862)
